requirements.txt
fastapi
uvicorn
tensorflow==2.12.0
Pillow
python-multipart
numpy
playwright
pandas